move data reader
