+++ sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/continuous/ContinuousQueuedDataReader.scala
@@ -26,6 +26,7 @@ import org.apache.spark.{Partition, SparkEnv, SparkException, TaskContext}
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.catalyst.expressions.UnsafeRow
 import org.apache.spark.sql.execution.datasources.v2.DataSourceRDDPartition
+import org.apache.spark.sql.sources.v2.reader.DataReader
 import org.apache.spark.sql.sources.v2.reader.streaming.PartitionOffset
 import org.apache.spark.util.ThreadUtils
 
@@ -109,6 +110,52 @@ class ContinuousQueuedDataReader(
   override def close(): Unit = {
     dataReaderThread.interrupt()
     epochMarkerExecutor.shutdown()
+  }
+
+  /**
+   * The data component of [[ContinuousQueuedDataReader]]. Pushes (row, offset) to the queue when
+   * a new row arrives to the [[DataReader]].
+   */
+  class DataReaderThread(
+      reader: DataReader[UnsafeRow],
+      queue: BlockingQueue[(UnsafeRow, PartitionOffset)],
+      context: TaskContext,
+      failedFlag: AtomicBoolean)
+    extends Thread(
+      s"continuous-reader--${context.partitionId()}--" +
+        s"${context.getLocalProperty(ContinuousExecution.EPOCH_COORDINATOR_ID_KEY)}") {
+    private[continuous] var failureReason: Throwable = _
+
+    override def run(): Unit = {
+      TaskContext.setTaskContext(context)
+      val baseReader = ContinuousDataSourceRDD.getBaseReader(reader)
+      try {
+        while (!context.isInterrupted && !context.isCompleted()) {
+          if (!reader.next()) {
+            // Check again, since reader.next() might have blocked through an incoming interrupt.
+            if (!context.isInterrupted && !context.isCompleted()) {
+              throw new IllegalStateException(
+                "Continuous reader reported no elements! Reader should have blocked waiting.")
+            } else {
+              return
+            }
+          }
+
+          queue.put((reader.get().copy(), baseReader.getOffset))
+        }
+      } catch {
+        case _: InterruptedException if context.isInterrupted() =>
+          // Continuous shutdown always involves an interrupt; do nothing and shut down quietly.
+
+        case t: Throwable =>
+          failureReason = t
+          failedFlag.set(true)
+          // Don't rethrow the exception in this thread. It's not needed, and the default Spark
+          // exception handler will kill the executor.
+      } finally {
+        reader.close()
+      }
+    }
   }
 
   /**
